+++ src/test/kotlin/world/cepi/kepi/singleton/Square.kt
@@ -1,6 +1,9 @@
-package world.cepi.kepi
+package world.cepi.kepi.singleton
 
 import kotlinx.serialization.Serializable
 
+/**
+ * Simple square model for non-id based data.
+ */
 @Serializable
 class Square(val width: Int, val height: Int)
