+++ src/test/kotlin/world/cepi/kepi/singleton/SquareModel.kt
@@ -1,10 +1,13 @@
-package world.cepi.kepi
+package world.cepi.kepi.singleton
 
 import kotlinx.serialization.encodeToString
 import world.cepi.kepi.data.DataNamespaceForge
 import world.cepi.kepi.data.Model
 import world.cepi.kepi.data.Model.Companion.jsonParser
 
+/**
+ * Singleton model -- storing a `Square` would only store this one instance of the square at the same ID.
+ */
 class SquareModel : Model<Square> {
 
     override val dataNamespace: DataNamespaceForge = DataNamespaceForge("shape", "sqare")
@@ -12,5 +15,7 @@ class SquareModel : Model<Square> {
     override fun asData(item: Square): Pair<String, String> {
         return "baseShape" to jsonParser.encodeToString(item)
     }
+
+    override val isSingleton = true
 
 }
